Got new collisions architecture compiling.

+++ source/objects/actors/MovingEntity/MovingEntity.h
@@ -103,7 +103,7 @@ void MovingEntity_setDefinition(MovingEntity this, void* movingEntityDefinition)
 void MovingEntity_ready(MovingEntity this, bool recursive);
 bool MovingEntity_handleMessage(MovingEntity this, Telegram telegram);
 void MovingEntity_setLocalPosition(MovingEntity this, const VBVec3D* position);
-int MovingEntity_getAxisFreeForMovement(MovingEntity this);
+u16 MovingEntity_getAxisFreeForMovement(MovingEntity this);
 void MovingEntity_startMovement(MovingEntity this);
 void MovingEntity_checkDisplacement(MovingEntity this);
